Replace UUIDS with unique_names (#18)
* removed UUIDS from Graph in borg

* replace convert_id calls with .name

* NameConverter class removed, move name assignment

* remove get_key

* Add check if object name is already taken

* add default name generation to BasedBase and Descriptor

* borg _clear method

* change id from name to unique_name

* more .name changes

* reorder name and value

* more .name replacements and test fixes

* Fix tests

* Ruff

* tox fix

* fix tox errors

* Update map on unique_name change + code cleanup

* Clear graph unittest

* test_add_vertex

* Rename borg to global_object

* Restructure sourcecode tree

* restructure test folder

* more tests

* test for identical unique names

* removed UUIDS from Graph in borg

* replace convert_id calls with .name

* NameConverter class removed, move name assignment

* remove get_key

* Add check if object name is already taken

* add default name generation to BasedBase and Descriptor

* borg _clear method

* change id from name to unique_name

* more .name changes

* reorder name and value

* more .name replacements and test fixes

* Fix tests

* Ruff

* tox fix

* Update map on unique_name change + code cleanup

* Clear graph unittest

* test_add_vertex

* Rename borg to global_object

* Restructure sourcecode tree

* restructure test folder

* more tests

* test for identical unique names

* more tests

* Rebase on develop

* fix dict tests

* fix __copy__ method

* fix tests

* fix remaining tests

* ruff & test syntax

* add final unique_name tests

* add borg alias for global_object for backwards compatibility

* According to PR

* fix mistake

* possible fix of test_star
