+++ examples_old/example4.py
@@ -9,7 +9,7 @@ from typing import List
 
 import numpy as np
 
-from easyscience import borg
+from easyscience import global_object
 from easyscience.Fitting import Fitter
 from easyscience.Objects.core import ComponentSerializer
 from easyscience.Objects.ObjectClasses import BaseObj
@@ -84,7 +84,7 @@ class InterfaceTemplate(ComponentSerializer, metaclass=ABCMeta):
     """
 
     _interfaces = []
-    _borg = borg
+    _global_object = global_object
 
     def __init_subclass__(cls, is_abstract: bool = False, **kwargs):
         """
@@ -171,7 +171,7 @@ class Interface1(InterfaceTemplate):
         :return: None
         :rtype: noneType
         """
-        if self._borg.debug:
+        if self._global_object.debug:
             print(f"Interface1: Value of {value_label} set to {value}")
         setattr(self.calculator, value_label, value)
 
@@ -224,7 +224,7 @@ class Interface2(InterfaceTemplate):
         :return: None
         :rtype: noneType
         """
-        if self._borg.debug:
+        if self._global_object.debug:
             print(f"Interface2: Value of {value_label} set to {value}")
         self._data = json.loads(self.calculator.export_data())
         if value_label in self._data.keys():
