+++ src/easyscience/Objects/new_variable/descriptor_base.py
@@ -8,9 +8,9 @@ import abc
 from typing import Any
 from typing import Optional
 
-from easyscience import borg
+from easyscience import global_object
+from easyscience.global_object.undo_redo import property_stack_deco
 from easyscience.Objects.core import ComponentSerializer
-from easyscience.Utils.UndoRedo import property_stack_deco
 
 
 class DescriptorBase(ComponentSerializer, metaclass=abc.ABCMeta):
@@ -27,11 +27,12 @@ class DescriptorBase(ComponentSerializer, metaclass=abc.ABCMeta):
     # Used by serializer
     _REDIRECT = {'parent': None}
 
-    _borg = borg
+    _global_object = global_object
 
     def __init__(
         self,
         name: str,
+        unique_name: Optional[str] = None,
         description: Optional[str] = None,
         url: Optional[str] = None,
         display_name: Optional[str] = None,
@@ -52,6 +53,11 @@ class DescriptorBase(ComponentSerializer, metaclass=abc.ABCMeta):
 
         .. note:: Undo/Redo functionality is implemented for the attributes `name` and `display name`.
         """
+
+        if unique_name is None:
+            unique_name = self._unique_name_generator()
+        self._unique_name = unique_name
+
         if not isinstance(name, str):
             raise TypeError('Name must be a string')
         self._name: str = name
@@ -74,10 +80,10 @@ class DescriptorBase(ComponentSerializer, metaclass=abc.ABCMeta):
 
         # Let the collective know we've been assimilated
         self._parent = parent
-        self._borg.map.add_vertex(self, obj_type='created')
+        self._global_object.map.add_vertex(self, obj_type='created')
         # Make the connection between self and parent
         if parent is not None:
-            self._borg.map.add_edge(parent, self)
+            self._global_object.map.add_edge(parent, self)
 
     @property
     def name(self) -> str:
@@ -165,6 +171,33 @@ class DescriptorBase(ComponentSerializer, metaclass=abc.ABCMeta):
         self._url = url
 
     @property
+    def unique_name(self) -> str:
+        """
+        Get the unique name of this object.
+
+        :return: Unique name of this object
+        """
+        return self._unique_name
+    
+    @unique_name.setter
+    def unique_name(self, new_unique_name: str):
+        """ Set a new unique name for the object. The old name is still kept in the map. 
+        
+        :param new_unique_name: New unique name for the object"""
+        if not isinstance(new_unique_name, str):
+            raise TypeError("Unique name has to be a string.")
+        self._unique_name = new_unique_name
+        self._global_object.map.add_vertex(self)
+
+    def _unique_name_generator(self) -> str:
+        """
+        Generate a generic unique name for the object using the class name and a global iterator.
+        """
+        class_name = self.__class__.__name__
+        iterator_string = str(self._global_object.map._get_name_iterator(class_name))
+        return class_name + "_" + iterator_string
+
+    @property
     @abc.abstractmethod
     def value(self) -> Any:
         """Get the value of the object."""
@@ -180,5 +213,7 @@ class DescriptorBase(ComponentSerializer, metaclass=abc.ABCMeta):
 
     def __copy__(self) -> DescriptorBase:
         """Return a copy of the object."""
-        new_obj = self.__class__.from_dict(self.as_dict())
+        temp = self.as_dict()
+        temp["unique_name"] = None
+        new_obj = self.__class__.from_dict(temp)
         return new_obj
