+++ src/easyscience/Objects/new_variable/descriptor_bool.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Any
 from typing import Optional
 
-from easyscience.Utils.UndoRedo import property_stack_deco
+from easyscience.global_object.undo_redo import property_stack_deco
 
 from .descriptor_base import DescriptorBase
 
@@ -16,6 +16,7 @@ class DescriptorBool(DescriptorBase):
         self,
         name: str,
         value: bool,
+        unique_name: Optional[str] = None,
         description: Optional[str] = None,
         url: Optional[str] = None,
         display_name: Optional[str] = None,
@@ -25,6 +26,7 @@ class DescriptorBool(DescriptorBase):
             raise ValueError(f'{value=} must be type bool')
         super().__init__(
             name=name,
+            unique_name=unique_name,
             description=description,
             url=url,
             display_name=display_name,
