+++ src/easyscience/__init__.py
@@ -4,13 +4,22 @@
 
 __author__ = 'github.com/wardsimon'
 
+import warnings
+
 import numpy as np  # noqa: F401  This is used in the other codebases that uses easyscience
 import pint
 
 from easyscience.__version__ import __version__ as __version__
-from easyscience.Objects.Borg import Borg
+from easyscience.global_object import GlobalObject
 
 ureg = pint.UnitRegistry()
-borg = Borg()
-borg.instantiate_stack()
-borg.stack.enabled = False
+global_object = GlobalObject()
+global_object.instantiate_stack()
+global_object.stack.enabled = False
+
+# alias for global_object, remove later
+def __getattr__(name):
+    if name == 'borg':
+        warnings.warn("The 'borg' has been renamed to 'global_object', this alias will be deprecated in the future", DeprecationWarning)  # noqa: E501
+        print("The 'borg' has been renamed to 'global_object', this alias will be deprecated in the future")
+        return global_object
