+++ src/easyscience/fitting/minimizers/minimizer_lmfit.py
@@ -16,7 +16,6 @@ from lmfit.model import ModelResult
 from .minimizer_base import MinimizerBase
 from .utils import FitError
 from .utils import FitResults
-from .utils import NameConverter
 
 
 class LMFit(MinimizerBase):  # noqa: S101
@@ -77,7 +76,7 @@ class LMFit(MinimizerBase):  # noqa: S101
         self._cached_pars = {}
         self._cached_pars_vals = {}
         for parameter in self._object.get_fit_parameters():
-            key = NameConverter().get_key(parameter)
+            key = parameter.unique_name
             self._cached_pars[key] = parameter
             self._cached_pars_vals[key] = (parameter.value, parameter.error)
 
@@ -95,7 +94,7 @@ class LMFit(MinimizerBase):  # noqa: S101
             # Update the `Parameter` values and the callback if needed
             # TODO THIS IS NOT THREAD SAFE :-(
             for name, value in kwargs.items():
-                par_name = int(name[1:])
+                par_name = name[1:]
                 if par_name in self._cached_pars.keys():
                     # This will take into account constraints
 
@@ -198,11 +197,11 @@ class LMFit(MinimizerBase):  # noqa: S101
             minimizer_kwargs = {'fit_kws': minimizer_kwargs}
         minimizer_kwargs.update(engine_kwargs)
 
-        # Why do we do this? Because a fitting template has to have borg instantiated outside pre-runtime
-        from easyscience import borg
+        # Why do we do this? Because a fitting template has to have global_object instantiated outside pre-runtime
+        from easyscience import global_object
 
-        stack_status = borg.stack.enabled
-        borg.stack.enabled = False
+        stack_status = global_object.stack.enabled
+        global_object.stack.enabled = False
 
         try:
             if model is None:
@@ -241,7 +240,7 @@ class LMFit(MinimizerBase):  # noqa: S101
         :rtype: LMParameter
         """
         return LMParameter(
-            'p' + str(NameConverter().get_key(obj)),
+            'p' + obj.unique_name,
             value=obj.raw_value,
             vary=not obj.fixed,
             min=obj.min,
@@ -258,15 +257,15 @@ class LMFit(MinimizerBase):  # noqa: S101
         :return: None
         :rtype: noneType
         """
-        from easyscience import borg
+        from easyscience import global_object
 
         pars = self._cached_pars
         if stack_status:
             for name in pars.keys():
                 pars[name].value = self._cached_pars_vals[name][0]
                 pars[name].error = self._cached_pars_vals[name][1]
-            borg.stack.enabled = True
-            borg.stack.beginMacro('Fitting routine')
+            global_object.stack.enabled = True
+            global_object.stack.beginMacro('Fitting routine')
         for name in pars.keys():
             pars[name].value = fit_result.params['p' + str(name)].value
             if fit_result.errorbars:
@@ -274,7 +273,7 @@ class LMFit(MinimizerBase):  # noqa: S101
             else:
                 pars[name].error = 0.0
         if stack_status:
-            borg.stack.endMacro()
+            global_object.stack.endMacro()
 
     def _gen_fit_results(self, fit_results: ModelResult, **kwargs) -> FitResults:
         """
