+++ src/easyscience/fitting/minimizers/utils.py
@@ -51,23 +51,6 @@ class FitResults:
     def reduced_chi(self):
         return self.chi2 / (len(self.x) - self.n_pars)
 
-
-class NameConverter:
-    def __init__(self):
-        from easyscience import borg
-
-        self._borg = borg
-
-    def get_name_from_key(self, item_key: int) -> str:
-        return getattr(self._borg.map.get_item_by_key(item_key), 'name', '')
-
-    def get_item_from_key(self, item_key: int) -> object:
-        return self._borg.map.get_item_by_key(item_key)
-
-    def get_key(self, item: object) -> int:
-        return self._borg.map.convert_id_to_key(item)
-
-
 class FitError(Exception):
     def __init__(self, e: Exception = None):
         self.e = e
