+++ src/easyscience/global_object/__init__.py
@@ -1,0 +1,4 @@
+from .global_object import GlobalObject  # noqa: F401
+from .hugger.hugger import ScriptManager  # noqa: F401
+from .logger import Logger  # noqa: F401
+from .map import Map  # noqa: F401
