+++ tests/unit_tests/Objects/new_variable/test_descriptor_base.py
@@ -1,6 +1,6 @@
 import pytest
 
-from easyscience import borg
+from easyscience import global_object
 from easyscience.Objects.new_variable.descriptor_base import DescriptorBase
 
 
@@ -10,7 +10,7 @@ class TestDesciptorBase:
         # This avoids the error: TypeError: Can't instantiate abstract class DescriptorBase with abstract methods __init__
         DescriptorBase.__abstractmethods__ = set()
         DescriptorBase.__repr__ = lambda x: "DescriptorBase"
-        self.objs_before_new_descriptor = len(borg.map.created_objs)
+        self.objs_before_new_descriptor = len(global_object.map.created_objs)
         descriptor = DescriptorBase(
             name="name",
             description="description",
@@ -19,6 +19,10 @@ class TestDesciptorBase:
             parent=None,
         )
         return descriptor
+
+    @pytest.fixture
+    def clear(self):
+        global_object.map._clear()
 
     @pytest.mark.parametrize("name", [1, True, 1.0, [], {}, (), None, object()], ids=["int", "bool", "float", "list", "dict", "tuple", "None", "object"])
     def test_init_name_type_error(self, name):
@@ -49,7 +53,7 @@ class TestDesciptorBase:
         assert descriptor._description == "description"
         assert descriptor._url == "url"
         assert descriptor._display_name == "display_name"
-        assert len(descriptor._borg.map.created_objs) - self.objs_before_new_descriptor == 1
+        assert len(descriptor._global_object.map.created_objs) - self.objs_before_new_descriptor == 1
 
 
     def test_display_name(self, descriptor: DescriptorBase):
@@ -112,18 +116,18 @@ class TestDesciptorBase:
 
 
     @pytest.mark.parametrize("stack_enabled,stack_elements", [(False, 0), (True, 1)])
-    def test_set_display_name_without_borg_stack(self, descriptor: DescriptorBase, stack_enabled, stack_elements):
+    def test_set_display_name_without_global_object_stack(self, descriptor: DescriptorBase, stack_enabled, stack_elements):
         # When
         descriptor.__repr__ = lambda x: "DescriptorBase"
-        descriptor._borg.stack.clear()
-        descriptor._borg.stack._enabled = stack_enabled
+        descriptor._global_object.stack.clear()
+        descriptor._global_object.stack._enabled = stack_enabled
 
         # Then 
         descriptor.display_name = "new_display_name" 
 
         # Expect
         assert descriptor.display_name == "new_display_name"
-        assert len(descriptor._borg.stack.history)  == stack_elements
+        assert len(descriptor._global_object.stack.history)  == stack_elements
 
     def test_copy(self, descriptor: DescriptorBase):
         # When Then
@@ -136,7 +140,7 @@ class TestDesciptorBase:
         assert descriptor_copy._url == descriptor._url
         assert descriptor_copy._display_name == descriptor._display_name
 
-    def test_as_data_dict(self, descriptor: DescriptorBase):
+    def test_as_data_dict(self, clear, descriptor: DescriptorBase):
         # When Then
         descriptor_dict = descriptor.as_data_dict()
 
@@ -146,4 +150,21 @@ class TestDesciptorBase:
             "description": "description",
             "url": "url",
             "display_name": "display_name",
+            "unique_name": "DescriptorBase_0",
         }
+
+    def test_unique_name_generator(self, clear, descriptor: DescriptorBase):
+        # When Then Expect
+        assert descriptor.unique_name == "DescriptorBase_0"
+
+    def test_unique_name_change(self, clear, descriptor: DescriptorBase):
+        # When Then
+        descriptor.unique_name = "test"
+        # Expect
+        assert descriptor.unique_name == "test"
+
+    @pytest.mark.parametrize("input", [2, 2.0, [2], {2}, None])
+    def test_unique_name_change_exception(self, input, descriptor: DescriptorBase):
+        # When Then Expect
+        with pytest.raises(TypeError):
+            descriptor.unique_name = input
