--- tests/unit_tests/Objects/test_graph.py
@@ -1,7 +1,0 @@
-__author__ = "github.com/wardsimon"
-__version__ = "0.1.0"
-
-#  SPDX-FileCopyrightText: 2023 EasyScience contributors  <core@easyscience.software>
-#  SPDX-License-Identifier: BSD-3-Clause
-#  © 2021-2023 Contributors to the EasyScience project <https://github.com/easyScience/EasyScience
-
